[BUGFIX] Avoid error message on unlimited upload_max_filesize
Resolves: #73214
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/47029
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/core/Classes/Resource/ResourceStorage.php
@@ -938,7 +938,7 @@ class ResourceStorage implements ResourceStorageInterface
         }
         // Max upload size (kb) for files.
         $maxUploadFileSize = GeneralUtility::getMaxUploadFileSize() * 1024;
-        if ($uploadedFileSize >= $maxUploadFileSize) {
+        if ($maxUploadFileSize > 0 && $uploadedFileSize >= $maxUploadFileSize) {
             unlink($localFilePath);
             throw new Exception\UploadSizeException('The uploaded file exceeds the size-limit of ' . $maxUploadFileSize . ' bytes', 1322110041);
         }
